Merge ij-platform into ij-core
The rationale for this change is to reduce the total number of modules,
to make the system easier to understand. The PlatformService is a core
component of the ImageJ application framework, which more easily enables
platform-specific extensions, while requiring no further dependencies;
as such, it makes sense for it to be available as part of ij-core.