Better handling of command line arguments
Rather than having support for command line arguments be a feature of
the UIService, we instead have a new ConsoleService that encapsulates
this logic.

There are some details that need to be kept in mind:

Some arguments may modify UI appearance or behavior, so the
ConsoleService must process the arguments before UIService.createUI() is
called.

In general, arguments that are specific to a particular UI, platform, or
other extension will need to be handled by that extension. So the
ConsoleService will merely cache such settings, rather than directly
altering other parts of the system in response.

To accomplish this, the ConsoleService will manipulate a ConsoleOptions
or other Options plugin(s) directly for the purposes of caching the
settings, so that downstream code does not need to care about the
ConsoleService specifically but rather only the options in question.