Add a script that populates or updates app/ImageJ.app/
This is much safer than the approach I chose earlier: the
copy-dependencies plugin unfortunately replaces -SNAPSHOT filenames with
concrete <date>-<build> ones, which I do not want. I really want basically
an unpacked ImageJ.app/.

So this script might make things a little slower but way more robust.

For compatibility with MacOSX -- which has BSD Unix tools with all its
limitations relative to the GNU Unix tools -- we do not use (GNU) xargs'
-d option but jump through a small hoop.

To make sure that it is in a state of being uploaded, you want to make
sure that your Eclipse is either closed or Project>Build Automatically...
is switched off. Also, it does not hurt to run "mvn clean" before this
script...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>