Use core plugin mechanism to discover services
This eliminates the @Service as a separate annotation type, in favor of
"@Plugin(type = IService.class)" to accomplish the same thing.

The difference is that now services are explicitly tracked by the
PluginIndex and PluginService as a specific type of plugin. It is
hopefully also less confusing for new developers, since services are no
longer a super special case, but rather merely an (important) example of
ImageJ's extensibility.