Eliminate unnecessary ImageJPlugin interface
It provided an extra layer on top of RunnablePlugin, so that it was
possible to define other types of RunnablePlugins that were not
ImageJPlugins. In particular, we were using it for the 3D Viewer to
define a Viewer3DPlugin type, so that it could construct its own menu
structure independent of the ImageJ menus. However, we now the menuRoot
@Plugin attribute to control that, so this extra interface is no longer
necessary.