Improve priority-related logic
The Prioritized interface now defines a setPriority(double) method to
explicitly set the priority. The service initialization mechanism uses
it to specify the priority of newly created services, based on
information received from the plugin index. This is a better approach
than hardcoding the service helper to glean the priority from the
@Plugin annotation directly.