Rename IJHotDrawOverlayAdapter to JHotDrawAdapter
This is a simpler name that is easier to read, and still conveys the
purpose for the most part. It is unlikely (famous last words!) that we
will have a *different* kind of adapter for some other entity within
JHotDraw, so hopefully dropping the "Overlay" from the name will be OK.

We also drop the "I" prefix because we are trying not to use the "I"
interface convention whenever possible.