JHotDraw adapters: rename *Adapter to Swing*Tool
This name is more in line with convention for Swing-specific
implementations of interfaces, and also hopefully more intuitive for
people looking for Tool implementations. (For example, if you are
looking for the rectangle tool, and search for *RectangleTool, you will
now find it, whereas before you would have needed to do a lookup on Tool
for all implementing classes and then browse through them.)

The "Swing" prefix is unfortunate, but it differentiates the
Swing-specific tool implementations from any other UI's implementation
(e.g., a possible future implementation of RectangleTool in pure AWT)
without needing to rely on the fully qualified package names.