Launcher (32-bit): disallow non-sensical memory settings in ImageJ.cfg
Especially in multi-platform settings, it is easy to slip in a value
for the memory setting that makes sense for 64-bit but not for 32-bit
mode. Just Do The Right Thing and cap the value in 32-bit mode just
like we would with options passed via the command line.

Requested by Curtis Rueden.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>