Updater (upload): ensure obsolete records are written
... even if another site serves the files now...

Take ij-launcher.jar for example: it used to be served from the Fiji
update site, but now it is served from the ImageJ update site. The ImageJ
update site has no notion (and should not need to have) of the previous
ij-launcher.jar versions Fiji had, so the Fiji update site still needs to
serve the obsolete record.

This fixes the updater in that the db.xml.gz written to the Fiji update
site upon upload no longer forgets the previous checksums. As a
consequence, the updater knows that it is okay to update ij-launcher.jar
versions which were installed from previous Fiji versions.

This should prevent Fiji bugs number 466 and 467 from happening in the
future.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>