Updater: work around StderrLogService bug
The bug was that the updater failed to update dependencies of the
Updater and the Service moved to ij-core.

To support people who do not update on a daily basis, let's have a very,
very minimal LogService.

Unfortunately, we need to use exactly the same name for that log service
as we did before, because the updater cannot successfully update itself
otherwise (we must assume that people have the updater which exposes the
bug that it updates *only* ij-updater-core, not ij-core nor
ij-ui-swing-updater).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>