Updater: put back the Canceled class to unbreak the updating process
The Canceled class was renamed to a better name, which is a sensible thing
to do.

However, since the updater only updates the updater core and no
UI-specific backend, this broke the Swing Updater rather badly: the core
was updated, so the Canceled class went poof. But the GUI was not updated,
so it went boom when started, trying to find the Canceled class which went
poof earlier.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>