Create InputWidget UIs after construction
This migrates the creation of UI-specific elements into a dedicated
InputWidget#initialize(WidgetModel) method, rather than doing it during
construction.

This change allows for instantiation of widget instances without
necessarily creating any UI elements. It will be useful later for
dynamic discovery and selection of widgets; we can create a widget,
inquire as to whether it is compatible with a given WidgetModel, and
only initialize its UI-specific objects if so.