Fix PlatformService#exec and Platform#open logic
Firstly, PlatformService#exec now returns the exit code rather than a
simple boolean. In the case of total execution failure, an IOException
is thrown anyway, so no harm done there.

Secondly, DefaultPlatform#open(URL) was calling an executable called
"open" and then if that was successful, it would move on to try a number
of other hardcoded browser executable. It seems to me we should omit the
execution of "open" in favor of simply trying every item on the
hardcoded list. In the case of OS X, it is handled differently in
MacOSXPlatform#open(URL) anyway (where it does indeed use the "open"
executable). Anyway, if using "open" preferentially on e.g. Linux is
preferable, we can just prepend "open" to the list of "browsers" to try.