Fix widget compatibility logic
The old logic was totally wrongheaded. This commit restores the logic to
be very similar to how it was before the change to extensible widgets.

One problem was the use of ClassUtils#canConvert to detect whether the
widget could handle a particular parameter type. For example, the
TextFieldWidgets were being erroneously chosen for many different
parameter types such as ColorRGB, because according to
ClassUtils#canConvert, any object can be trivially converted to a String
using Object#toString().

Really, a TextFieldWidget should only be chosen if the type actually
*is* a String (or a char or a Character), and not merely because the
parameter type could be converted to a String.

Another problem was the implementation for choice and text field
widgets; the former should only claim compatibility if the parameter
type is a string, while the latter should only claim compatibility of
the parameter is not multiple choice.