Reimplement support for messages as parameters
There is now a MessageWidget that is used in the case of MESSAGE
visibility parameters.

This commit introduces two new flags for widgets: isLabeled and
isMessage. If isLabeled is false, the label prefix is omitted and
instead the widget spans the entire dialog. This could be useful beyond
just messages, for things like graphs, charts or other embedded UI.
If isMessage is true, then the widget counts as a message-only widget,
which can affect how the input dialog is rendered. In particular, if
there are only message widgets, then the dialog may show a "question" or
"information" icon; see SwingInputHarvester for details.