Fix some style issues
In particular, the preferred order of code in an ImageJ class is:

1. Static fields.
2. Instance fields.
3. Constructors.
4. New public methods for that class specifically.
5. Public method overrides of the class's superclass(es).
6. Public method implementations of the class's implemented
   interface(s), in the same order they appear in the interface(s).
7. Protected event handler methods (labeled "-- Event handlers --").
8. Private methods (labeled "-- Helper methods --").

Personally I find it very confusing when a method is out of position. It
is easy to assume that a particular method is not overridden in a given
class if the method does not appear in the same position as the
superclass and/or interface in question.

Similarly, if an instance field is mixed in with method declarations
later in the class, it can be difficult to realize that the variable in
question is in fact an instance field of that class (rather than e.g. a
protected member of a superclass).