JHotDrawImageCanvas: remove serialVersionUID
Including a serialVersionUID constant is an implicit promise that we are
paying attention to the class's serialization capabilities. Other
developers might assume that because there is a version ID given, they
can safely serialize objects of the class. In truth, we do not intend
for JHotDrawImageCanvas objects to ever be serialized, so it seems more
honest to leave off the serialVersionUID entirely.