JHotDrawImageCanvas: clarify event handling
Most of the events handled by the JHotDrawImageCanvas are only relevant
when the event pertains to *that* canvas. This changes makes it
hopefully easier to understand exactly when that condition holds. It
also reduces the amount of nested indentation, for readability, and
unifies the event object's variable name as "event" in all cases.