Migrate DisplayViewer classes into ij-ui
The DisplayViewer is a particular type of ImageJ plugin intended to
connect a Display (which is UI-agnostic) with a particular UI. So there
are implementations in the various ImageJ UI components (e.g.,
SwingImageDisplayViewer in ij-ui-swing-base). Hence, it makes most sense
for the base DisplayViewer classes to live in the base ij-ui component.