InputWidget: use simpler imagej.widget package
This is less horrible to read than "imagej.ext.module.ui" and also less
confusing since the "ui" implies something to do with the ij-ui
hierarchy, which is not the case here.

Rather, InputWidget is merely one type of ImageJ plugin, similar to
Service in imagej.service or Platform in imagej.platform.