Merge ij-ext-swing into ij-ui-swing-base
This change is part of an ongoing effort to simplify the ImageJ
component structure to make it easier to understand.

The difference between ij-ext-swing and ij-ui-swing-base is that the
former provided Swing implementations of interfaces from ij-core, while
the latter provided Swing implementations of interfaces from ij-ui.
Hence, the main downside of the merger is that ij-ui is now a required
dependency just to use the Swing widgets. But I think it is worth it so
that more Swing code is located in a single JAR file.