Merge ij-ext-swt into ij-ui-swt
This change is part of an ongoing effort to simplify the ImageJ
component structure to make it easier to understand.

The difference between ij-ext-swt and ij-ui-swt is that the former
provided SWT implementations of interfaces from ij-core, while the
latter provided SWT implementations of interfaces from ij-ui. Hence, the
main downside of the merger is that ij-ui is now a required dependency
just to use the SWT widgets. But I think it is worth it so that all SWT
code is located in a single JAR file.