Merge ij-ext-pivot into ij-ui-pivot
This change is part of an ongoing effort to simplify the ImageJ
component structure to make it easier to understand.

The difference between ij-ext-pivot and ij-ui-pivot is that the former
provided Pivot implementations of interfaces from ij-core, while the
latter provided Pivot implementations of interfaces from ij-ui. Hence,
the main downside of the merger is that ij-ui is now a required
dependency just to use the Pivot widgets. But I think it is worth it so
that all Pivot code is located in a single JAR file.