Merge ij-ext-awt into ij-ui-awt
This change is part of an ongoing effort to simplify the ImageJ
component structure to make it easier to understand.

The difference between ij-ext-awt and ij-ui-awt is that the former
provided AWT implementations of interfaces from ij-core, while the
latter provided AWT implementations of interfaces from ij-ui. Hence, the
main downside of the merger is that ij-ui is now a required dependency
just to use the AWT widgets. But I think it is worth it so that more AWT
code is located in a single JAR file.