Merge Swing SDI and MDI user interface components
This combines ij-ui-swing-sdi and ij-ui-swing-mdi components into
ij-ui-swing (formerly ij-ui-swing-base). The rationale is that there
were a large number of confusing Swing-related components. The number is
still rather high, but at least bundling SDI and MDI together into a
single "Swing UI" component is simpler than before.

The tricky part will be controlling under what conditions the SDI vs.
MDI UI gets instantiated and used when UIService#createUI() is called.
The plan is for the MDI interface to have lower priority than the SDI
one, so by default the SDI interface appears. But there will need to be
a mechanism to explicitly specify which UI gets used.