Rebroadcast only key & mouse events as appropriate
We err on the side of thoroughness when rebroadcasting key events,
because we do not know which component will have the keyboard focus at
any given time (and in some cases, the focus component is null, so the
window itself is responsible for handling the key events).

When it comes to mouse events, we must be more cautious. We do not want
mouse events from any old UI component being rebroadcast, but only from
the image canvas proper. This is because tools like the pixel probe want
to respond to mouse motion events as though those coordinates are within
the image canvas specifically.

These changes fix event propagation issues brought to light by the
recent Swing SDI/MDI merger. For some mysterious reason, these bugs did
not manifest prior to the merger, but at least some of them were
definitely latent in the codebase beforehand.