Rename imagej.ext.menu to imagej.menu
This is part of an effort to make the ImageJ package structure simpler
and easier to understand. The "imagej.ext" package hierarchy made sense
when the ImageJ extensibility framework was separated into its own
component (ij-ext), but now that it has merged with ij-core it is
confusing. It is also inconsistent which types of plugins were
subpackaged off of imagej.ext versus imagej (e.g., imagej.ext.display
for Displays versus imagej.platform for Platforms).