Migrate imagej.ext classes to imagej base package
(We leave only InstantiableException, temporarily, to avoid an issue
with old versions of the ImageJ updater.)

This is part of an effort to make the ImageJ package structure simpler
and easier to understand. The "imagej.ext" package hierarchy made sense
when the ImageJ extensibility framework as separated into its own
component (ij-ext), but now that it has merged with ij-core it is
confusing. It is also inconsistent which types of plugins were
subpackaged off of imagej.ext versus imagej (e.g., imagej.ext.display
for Displays versus imagej.platform for Platforms).

In the case of the imagej.ext package itself (not its subpackages), the
classes there are used in various places throughout the ImageJ codebase,
and are generally well situated in the base imagej package itself.