Rename imagej.ext.plugin to imagej.plugin
This is part of an effort to make the ImageJ package structure simpler
and easier to understand. The "imagej.ext" package hierarchy made sense
when the ImageJ extensibility framework as separated into its own
component (ij-ext), but now that it has merged with ij-core it is
confusing. It is also inconsistent which types of plugins are
subpackaged off of imagej.ext versus imagej (e.g., imagej.ext.display
for Displays versus imagej.platform for Platforms).

Unfortunately, due to a limitation in the ImageJ updater (which is fixed
in recent versions), we must leave a few classes in imagej.ext.plugin
for now. Specifically, those classes on which the updater code depends
directly must stay in the same place, since old versions of the updater
do not update their dependencies when updating themselves. Hence, these
updaters would pull down a new ij-updater without pulling down the
corresponding new ij-core, thus breaking that system's installation.