UIService: add more createUI methods
These methods allow creation of a specific UI, rather than the default
one. The UI can be specified via its name or class name using the
createUI(String) method, or by an actual instance using the
createUI(UserInterface) method -- even if that UserInterface was not
among those discovered automatically.

We also respect the value of the ij.ui system property now, creating the
UI named in that property by default (if given). If the property is not
set, the default UI is still the one with the highest priority.