Simplify ij-app profile configuration
Rather than having a separate profile for each alternative UI, we now
have a single profile called alt-ui that is activated whenever the ij.ui
system property is set. It merely adds all alternative UI component JARs
to the runtime classpath. The correct UI is now instantiated by the
UIService based on the value of that system property.