Add interface for objects housed by a UI
This UIComponent interface is shared between InputPanel and InputWidget,
and could potentially be useful for other composition-style UI (i.e,
"has a" UI component rather than "is" one) in the future.

One big advantage of composition-style UI is that you need not actually
*construct* said UI when the object is first instantiated; rather, it
can be done lazily, when it is time for the object to be
displayed/rendered onscreen.