SWTInputHarvester: get rid of static Display
The static SWT Display object can cause big problems. First of all, it
was wrong anyway; we should be using the same Display as the associated
SWTUI. Secondly, in the case of a different default UI, the
SWTInputHarvester would still try to create the static Display, which
does not work if native SWT libraries are not in the Java library path.

So now we obtain the Display as needed from the linked SWTUI, which
should always be the default display if we get far enough along in the
SWTInputHarvester execution to actually need it.