PivotUI: construct Pivot application differently
This change fixes many problems with the way the Pivot UI was previously
handled. The new approach is to spawn the PivotApplication context,
forcibly extract it from Pivot's static DesktopApplicationContext, then
do the actual initialization (including injection of the ImageJ context)
after the fact.

Since we then have a handle on the PivotApplication, we can finish
implementing the accessor methods for the PivotApplicationFrame,
PivotToolBar and PivotStatusBar.