PivotInputHarvester: improve harvestInputs UI
Previously, the Sheet was opened with a null owner, which is actually
illegal. So it had no chance of working. Now that the PivotUI provides
access to its PivotApplicationFrame (which is a Pivot Frame), we can use
that for the owner.

Unfortunately, the sheet still does not render properly (it appears as a
tiny white box), but we are getting closer...