Report status for rectangle, overlay & line tools
This feature necessitated the addition of some extra hooks into the
JHotDrawAdapter, which the SwingRectangleTool, SwingOverlayTool and
SwingLineTool then leverage to produce the status updates.

There is still a bug when moving the first endpoint of an existing line,
where the (X, Y) coordinates are given as the second endpoint instead,
rather than reporting the (X, Y) coordinates the endpoint being updated,
like ImageJ1 does.