ServiceHelper: do not add services whose classes cannot be loaded
If -- for whatever reason -- a given Service's class cannot be loaded,
it is quite safe to assume that the Service will not run, either. So
it does not make sense to add it in such a case.

This happens for example if somebody has a borked Fiji installation
where eventbus is missing while .jar files are present which use
the eventbus.

To be safe, let's not just handle a ClassNotFoundException but all
kinds of Throwables.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>