ClassUtils: add getLocation utility method
This method returns the classpath component containing the class with
the given name.

If the class is directly on the file system (e.g.,
"/path/to/my/package/MyClass.class") then it will return the base
directory (e.g., "/path/to").

If the class is within a JAR file (e.g.,
"/path/to/my-jar.jar!/my/package/MyClass.class") then it will return the
path to the JAR (e.g., "/path/to/my-jar.jar").