Move FileUtils#getBaseDirectory to AppUtils
These methods make more sense in a separate utility class from
FileUtils.

To avoid breaking the Updater, the old method signatures remain
(deprecated) in FileUtils, delegating to the AppUtils versions.

As part of the move, the implementations have been revamped. Hopefully
they are now clearer, more general, and more targeted where warranted.
There is no longer a reliance on locating pom.xml files to determine the
application execution environment, but rather a hint is given to
getBaseDirectory in the form of an expected subdirectory structure.