Revert "Add API to execute an existing plugin instance"
This reverts commit 399d9d379ea74745e818c4207e36ce1b2133d862.

It removes the ability to call PluginService#run on an existing
RunnablePlugin instance. Some discussion is present in ticket #1444
(http://trac.imagej.net/ticket/1444), but essentially, the problem is
that it is too easy to get confused about the state of the
RunnablePlugin instance. The instance itself has no way of tracking
which of its parameters should be considered "resolved" for the purposes
of module execution. So even if you call myPlugin.setXYZ(xyz), the
plugin service has no way of knowing you intended for xyz to be
finalized with that value, and plugin preprocessors may (probably will)
still overwrite its value.