Change behavior of ModuleService#run(?, Object...)
These methods previously took a list of input parameter values, in the
same declaration order as the module. This commit changes it to expect a
list of name/value pairs, which is much more resistant to future module
code changes. That is, if a module adds, removes or renames an input
parameter, or reorders the parameters, this new approach will tolerate
that much better because all inputs are explicitly named in the list.

In the case of a renamed or deleted parameter, a warning will be issued
stating that the old name is invalid, but at least there will be a
visible message in the log about it now.