Remove log messages from ClassUtils
This avoids the issue of the log being a contextual service now.

It is also not that big of a deal since in almost every case (except for
setValue(Field, Object, Object)), the method provides a return value
indicating what went wrong (e.g., loadClass returns null if the class
could not be loaded). In the case of setField, we now throw
IllegalArgumentException (an unchecked exception) if the field cannot be
set due to illegal access.