Add ImageJ#getInfo(boolean) method
This is a UI agnostic version of the info string previously computed by
SwingStatusBar#getInfoString(boolean) method. It has nothing to do with
Swing and is better as a method of the contex itself, so that other
areas of the code can make use of it, too.

fix: ImageJ#getInfo()