Add plugin base class for easier plugin invocation
The base class is a RunnablePlugin that is Contextual. Calling
setContext automatically populates service parameters, so it can be run
programmatically more easily (i.e., without using PluginService#run).
This approach is a more type-safe way of invoking a plugin.

The (small) downside is that plugins must extend the ContextPlugin base
class, rather than merely implementing the RunnablePlugin interface.