Rename RunnablePlugin to Command
This has ripples throughout the codebase, but is hopefully a much better
name, more concise and intuitive, describing what this type of plugin is
for: executing a command.

The term "command" is familiar to users of ImageJ1, the only difference
being that in ImageJ1, a "plugin" could implement multiple "commands",
but in ImageJ2 you would create a separate class for each desired
command anyway, so the distinction has been dropped.