Eliminate redundant options plugin constructors
Now that DynamicCommand is contextual, it makes the most sense to load
values from persistent storage when the context is set for each
particular options plugin. This ensures that field initializations have
already run (since the constructor is already finished by then).

It also fixes an NPE caused by the recent addition of CommandService.