Rename AbstractAssignPlugin to simpler MathCommand
The new name accurately and simply conveys the purpose of the class. Of
course, it does not provide utter detail (e.g., from the name, you
cannot tell that it implements Previewable), but a cursory inspection of
the class hierarchy will show that it does. And since there is no class
of non-previewable math commands with which it would be contrasted, that
detail is a less crucial one.