ObjectIndex: allow duplicate items
The object index uses lists internally, and the restriction that each
item could only be added to the index once was unnecessary and perhaps
counterintuitive. Certainly it impacted performance negatively, since
the lists had to be checked every time for the item's presence.

Within ImageJ's current uses of ObjectIndex, it should not be the
case that any item appears more than once anyway, so this change is
unlikely to impact ImageJ's behavior directly. There may be a situation
in the future where it is nice to have duplicate items. And in the case
where duplicate items are *not* desired, downstream code can check
itself for them using contains before adding to the index.