Create IJ1 Rois before running the first legacy command
We need to take care to make sure that everything is ready for the legacy
command before it runs. That includes the Roi (if any), even if we just
fired up the first legacy command of this session (i.e. even if the
harmonizer was just initialized and had no ImagePlus for any ImageDisplay
yet).

While at it, consolidate duplicated code into a common code path.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>