PluginService: add createInstance(PluginInfo<P>)
This method creates an instance of a single plugin. In the past this was
considered trivial by calling info.createInstance() but it is
increasingly necessary in many cases to then inject a context and a
priority into plugin types that implement Contextual and Prioritized,
respectively. So the plugin service now provides this capability.