ModuleService: add createModule(ModuleInfo)
This method creates an instance of a single module. In the past this was
considered trivial by calling info.createModule() but it is increasingly
necessary in many cases to then inject a context and a priority into
modules that implement Contextual and Prioritized, respectively. So the
module service now provides this capability.