PluginService: improve getPluginsOfClass(Class)
Specifically, the performance should be better now. We use the given
Class reference to look up the plugin type as given by the @Plugin
annotation, and then limit our search to PluginInfos of that type only.

Note that this is now feasible since we recently switched the @Plugin
annotation to a RUNTIME retention (it was COMPILE time before, meaning
that javac would erase it after running SezPoz on it).