OptionsService: instantiate options more robustly
We now always lean on the internal DefaultOptionsService#createInstance
method for creating instances of OptionsPlugins. This ensures they are
completely populated with their context, priority and service
parameters.

Because we do this, we also no longer need to run the
ServicePreprocessor on them explicitly afterwards, because their
services will already be populated.

Lastly, we invoke the remaining desired preprocessor plugins explicitly
rather than via a ModuleRunner, because we do not want any events to be
issued with respect to those preprocessing steps; the processing is
merely internal to how OptionsPlugins work.