Add miscellaneous utility class
We avoided it for a long time, but I am too uncreative (lazy?) right now
to think of a better place to put the initial methods of this class:
some helper methods for comparing two objects when one or both of them
might be null. So for the time being, MiscUtils it is!