Allow buttons with actions as @Parameter in plugins
This version is an initial experimental implementation. It only has
a Swing version and would need equivalents in other UIs. It is tested
in the Process > Math > Equation plugin.

Using this facility one can embed a button in a parameter dialog. When the
button is pressed the callback of the @Parameter is invoked. The callback
can then do any kind of processing it likes.