Register the SwingImageWindow only once to listen to key events
In commit d257602(Rebroadcast only key & mouse events as appropriate), the
event handling was changed slightly to guarantee that events are handled
only once. However, the SwingImageWindow was registered to listen for
keyboard events not only in SwingUI's createDisplayWindow(Display) method
-- when the SwingDisplayWindow was created -- but also in
SwingSdiImageDisplayViewer's view(DisplayWindow, Display) method.

We simply remove the latter one so that all SwingImageWindows get
registered, but only once.

Annoyingly, starting ImageJ2 in Eclipse did not trigger the error, because
the order of DisplayViewers (as found by
DefaultUIService#onEvent(DisplayCreatedEvent)) is not stable: Eclipse
found them in this order: SwingMdiImageDisplayViewer,
SwingTextDisplayViewer, SwingSdiImageDisplayViewer. Since the first viewer
already handled the display, the SDI implementation was never asked to
show the window.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>