DisplayTest: fix to work without DefaultDisplay
We made DefaultDisplay undiscoverable for the time being, but
DisplayTest relied on it, creating displays for some double values.

For the moment, we use strings instead of doubles, and all the tests
still pass. In the future, we will have a default display viewer, and
will then be able to reenable the use of doubles in these tests.