Updater: install dependencies even of up-to-date files
In the unfortunate case that a dependency was not marked correctly as
such, we would like to upload a fixed db.xml.gz. The updater needs to pick
that up.

Concrete example: if loci-common.jar depends on slf4j.jar, but slf4j.jar
was not listed correctly as a dependency, users could end up having
loci-common without slf4j. Uploading a fixed db.xml.gz would not fix it
for them. Until this commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>