Fix how displays are named
It is the responsibility of each individual Display implementation to
name itself in a reasonable fashion. Often this will somehow reflect the
objects that it displays. For example, an ImageDisplay will name itself
based on its constituent Data objects, which implement the
net.imglib2.meta.Named interface.

That said, we also support two other potentially common situations:

1. Downstream code wishes to override the name explicitly when
   displaying an object. That case is covered by the method
   DisplayService#createDisplay(String, Object).

2. The display really has no way to pick a good name from the displayed
	 object(s), and so leaves its name null or empty. In this case, if the
	 display is being used for an output parameter of a command execution,
	 the DisplayPostprocessor will have a better hint as to a suitable
	 name: the label (or failing that, the name) of the associated
	 ModuleItem. But we only fall back to those if the display is still
	 nameless after being passed the output value.