Code Hacker: require a class loader in which to define the classes
This opens the door to have multiple LegacyServices, each with their
own class loader, in which the ImageJ 1.x classes and the classes
using them are defined.

This is necessary because ij.IJ is basically a singleton but still has
to be connected to the appropriate ImageJ context and LegacyService. By
having a separate class loader for each LegacyService, ij.IJ can be
a singleton *in that class loader*, and things will still work out.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>