Add Fiji's extensible primitive type arrays
In Fiji (actually, in the Virtual Insect Brain project), we needed
primitive arrays that are more performant than the boxed
ArrayList<Double> (actually, in the VIB project, we did not have Java 5
yet, so we required the implementation anyway).

This developer implemented the classes from scratch, based on ideas
floating about on the net, hence it is no problem to re-license them to
the BSD license as required by ImageJ2.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>