Add resizable array list implementation
This code is necessary because there is no built-in public way to resize
an ArrayList. You can request the size with size(), add or remove
elements one at a time, or in batches, but cannot simply say "set the
size to X" while leaving any new elements as null, or efficiently
purging old extra elements.

For removal, the best you could do with the public API would be
repeatedly calling remove(size() - 1). However, ArrayList has a
protected removeRange(int, int) method that does the trick.

For addition, you can call ensureCapacity(int) to make sure the internal
array is large enough, but then there is absolutely no way to update the
size field; it is private. So we must resort to reflection.