UpdaterTest: do not trust .checksums
We're writing out files way too fast for .checksums to be correct all
the time (it checksums and timestamps, the latter with whole-second
precision, so it is highly likely that our test writes out files
which are then not picked up as modified.

We already handled that problem by deleting the .checksums file before
running the checksummer, but in the process of introducing
FilesCollection#downloadIndexAndChecksum(Progress) a conditional case was
introduced where this developer forgot to add the .checksums deletion
again.

This commit fixes that oversight.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>