Script Editor: implement search & replace
The Fiji Script Editor relied on RSyntaxTextArea's search & replace,
but we tried to be independent of that component, as RSyntaxTextArea
was LGPLed at the time. In the meantime, RSyntaxTextArea is
BSD-licensed, but let's just complete a rudimentary version of the
script editor that does not rely on any 3rd-party components.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>