PluginService: fix bug in getPluginsOfClass(Class)
It assumed every plugin is marked by an @Plugin annotation, which is of
course not the case (e.g., LegacyCommands are not annotated, but rather
are programmatically generated by the LegacyPluginFinder).

The fix is to fall back to search the entire list of ImageJPlugins in
the case where an @Plugin annotation is not found. There is surely a
more clever efficient solution, but this approach works for now.