Type change can now combine channels
This is a somewhat more intuitive way to make grayscale and color
images from multichannel input data. It now asks if you want to
combine channels during type change. The old way depended upon the
current setting of the Channels Tool which was cumbersome.

Right now it always asks even if the dataset does not have a channels
axis. We need to make it so that it only asks when data has more than
one channel. This could likely be supported with a DynamicPlugin but
some investigation is needed.