Add sandbox command showing param assignment order
With all the ways an input parameter can be assigned a value, it is nice
to have a little example that illustrates when each sort of assignment
occurs. This command dumps the parameter values to the log (or stderr if
the log is not yet available) at each step of the way.