Fix ClassUtils#getLocation method
1) It now returns a URL instead of a File. The rationale is that
sometimes a class comes from a resource other than a file, but that
resource can always be expressed as a URL, as given by both:
  c.getResource(c.getName() + ".class")
and
  c.getProtectionDomain().getCodeSource().getLocation()

2) We first try the protection-domain-based route to obtaining the
class's base location, since it works within the OSGi framework.
Conversely, when used within OSGi, the getResource route does not
provide a sensible URL from which a file can be derived, but rather
gives a URL starting with "bundleresource://".