Updater: fix ImageJ base directory detection
This change partially reverts commit
226a9f56cf9f1eb7dd41e6b73ea3898470a9d089.

With the preceding improvements to AppUtils.getBaseDirectory(), the
no-args signature is now robust enough to be called in this situation.

The rationale is that explicitly asking for the base directory with
respect to imagej.updater.util.Util.class will fail in the case where
that class is contained in a JAR file in the Maven local repository
cache. This is in fact the case when launching ImageJ via Maven
("mvn exec:exec" from the app subdirectory). Unfortunately, if the
Updater's Util class fails to detect the ImageJ base directory, its
static initializer throws an NPE, which means the class does not load.
This wreaks all sorts of havoc, and ImageJ fails to start up properly.