Do not try to handle keyboard shortcuts with key code 0
When an accelerator could not be parsed, its corresponding KeyCode is
set to UNDEFINED.

When a keyboard event was received for a key without an associated key
code, say, an umlaut, the corresponding KeyCode is UNDEFINED, too.

However, pressing a key without a key code does not mean that we should
run the first command that has an invalid accelerator.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>